requirements.txt
alt-profanity-check==0.24.0
better-profanity==0.7.0
blis==0.7.4
catalogue==1.0.0
certifi==2020.11.8
chardet==3.0.4
cymem==2.0.5
decorator==4.4.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
joblib==1.0.1
moviepy==1.0.3
murmurhash==1.0.5
numpy==1.19.3
oauthlib==3.1.0
Pillow==8.0.1
pip-autoremove==0.9.1
plac==1.1.3
preshed==3.0.5
proglog==0.1.9
PySocks==1.7.1
python-dotenv==0.19.0
requests==2.25.0
requests-oauthlib==1.3.0
scikit-learn==0.24.0
scipy==1.6.3
six==1.15.0
spacy==2.3.4
srsly==1.0.5
thinc==7.4.4
threadpoolctl==2.1.0
tqdm==4.54.1
tweepy==3.9.0
urllib3==1.26.2
wasabi==0.8.0
